# docs/swagger/swagger.yaml
basePath: /api
definitions:
...
  dtos.CharacterCreateDTO:
    properties:
      category_ids:
        items:
          type: integer
        type: array
      description:
        maxLength: 70
        type: string
      name:
        maxLength: 20
        type: string
      price:
        type: integer
      rarity_id:
        type: integer
    required:
    - description
    - name
    - price
    - rarity_id
    type: object
  dtos.CharacterReadDTO:
    properties:
      categories:
        items:
          type: string
        type: array
      description:
        type: string
      id:
        type: integer
      img_uri:
        type: string
      name:
        type: string
      price:
        type: integer
      rarity:
        type: string
    type: object
  dtos.CharacterUpdateDTO:
    properties:
      category_ids:
        items:
          type: integer
        type: array
      description:
        maxLength: 70
        type: string
      id:
        type: integer
      name:
        maxLength: 20
        type: string
      price:
        type: integer
      rarity_id:
        type: integer
    required:
    - description
    - id
    - name
    - price
    - rarity_id
    type: object
  dtos.ErrResp:
    properties:
      error:
        type: string
      path:
        type: string
      timestamp:
        type: string
    type: object
...
  dtos.TopCharactersDTO:
    properties:
      categories:
        items:
          type: string
        type: array
      count_of_purchases:
        type: integer
      description:
        type: string
      id:
        type: integer
      img_uri:
        type: string
      name:
        type: string
      price:
        type: integer
      rarity:
        type: string
    type: object
...
info:
  contact: {}
  description: This site documents the endpoints for the smaash web app, allowing
    easy testing
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: SMAASH API documentation
  version: "1.0"
paths:
...
  /characters:
    get:
      consumes:
      - application/json
      description: Reads characters
      produces:
      - application/json
      responses:
        "200":
          description: returns all characters
          schema:
            items:
              $ref: '#/definitions/dtos.CharacterReadDTO'
            type: array
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/dtos.ErrResp'
      tags:
      - characters
    post:
      consumes:
      - application/json
      description: Creates a new character
      parameters:
      - description: dto for creating a new character
        in: body
        name: character_create_dto
        required: true
        schema:
          $ref: '#/definitions/dtos.CharacterCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: returns newly created character
          schema:
            $ref: '#/definitions/dtos.CharacterReadDTO'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "409":
          description: unique key violation
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "422":
          description: request body in wrong format
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/dtos.ErrResp'
      tags:
      - characters
  /characters/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a character with the given id
      parameters:
      - description: id of desired character
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: doesn't return anything
          schema:
            type: ""
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "404":
          description: record not found
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/dtos.ErrResp'
      tags:
      - characters
    get:
      consumes:
      - application/json
      description: Reads a character by it's id
      parameters:
      - description: the id of the desired character
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: returns the desired character
          schema:
            $ref: '#/definitions/dtos.CharacterReadDTO'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "404":
          description: record not found
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/dtos.ErrResp'
      tags:
      - characters
    put:
      consumes:
      - application/json
      description: Updates the character with the given id
      parameters:
      - description: dto for updating a character
        in: body
        name: character_update_dto
        required: true
        schema:
          $ref: '#/definitions/dtos.CharacterUpdateDTO'
      - description: id of desired character
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: doesn't return anything
          schema:
            type: ""
        "400":
          description: id from url and id from request body doesn't match
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "404":
          description: record not found
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "409":
          description: unique key violation
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "422":
          description: request body in wrong format
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/dtos.ErrResp'
      tags:
      - characters
  /characters/{id}/img:
    get:
      consumes:
      - application/json
      description: Returns an uploaded character image
      parameters:
      - description: id of desired character
        in: path
        name: id
        required: true
        type: integer
      produces:
      - multipart/form-data
      responses:
        "404":
          description: record not found
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/dtos.ErrResp'
      tags:
      - characters
    post:
      consumes:
      - multipart/form-data
      description: Uploads an image for the character with given id
      parameters:
      - description: id of desired character
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: returns newly created image's URI
          schema:
            type: string
        "400":
          description: no file sent
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "415":
          description: invalid media type
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/dtos.ErrResp'
      tags:
      - characters
...
  /stats/top/characters:
    get:
      consumes:
      - application/json
      description: Reads the most popular items
      produces:
      - application/json
      responses:
        "200":
          description: returns items ordered by count of purchsases
          schema:
            items:
              $ref: '#/definitions/dtos.TopCharactersDTO'
            type: array
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/dtos.ErrResp'
      tags:
      - stats
...
swagger: "2.0"